+++ src/components/Button.vue
@@ -1,5 +1,5 @@
 <template>
-  <button @click="onClick" :class="styles" :disabled="disabled">
+  <button @click="onClick" :class="styles" :disabled="disabled" class="outline-none">
     <slot></slot>
   </button>
 </template>