+++ src/store/modules/sdk.ts
@@ -56,11 +56,13 @@ export interface SendData {
 export interface SDKState {
   balance: BigNumber | null
   sending: boolean
+  blacklist: Set<number>
 }
 
 const state: SDKState = {
   balance: null,
   sending: false,
+  blacklist: new Set(),
 }
 
 const mutations = <MutationTree<SDKState>>{
@@ -73,10 +75,15 @@ const mutations = <MutationTree<SDKState>>{
     console.log('{dispatch} transaction send in process: ', sending)
     state.sending = sending
   },
+
+  [types.SET_BLACKLIST](state: SDKState, blacklist: Set<number>) {
+    console.log('{dispatch} transaction send in process: ', blacklist)
+    state.blacklist = blacklist
+  }
 }
 
 const actions = <ActionTree<SDKState, RootState>>{
-  instantiateNomad() {
+  async instantiateNomad({ dispatch }) {
     console.log('instantiateNomad: ', environment)
     try {
       nomad = _instantiateNomad()
@@ -85,6 +92,13 @@ const actions = <ActionTree<SDKState, RootState>>{
       console.error(e)
       throw new Error("Couldn't setup Nomad")
     }
+    await dispatch('checkFailedHomes')
+  },
+
+  async checkFailedHomes({ commit }) {
+    await nomad.checkHomes(Object.keys(networks))
+    const blacklist = nomad.blacklist()
+    commit(types.SET_BLACKLIST, blacklist)
   },
 
   async getBalanceFromWallet({ rootState, commit }) {
@@ -163,7 +177,7 @@ const actions = <ActionTree<SDKState, RootState>>{
     nomad.registerSigner(networkName, newSigner)
   },
 
-  async send({ commit }, payload: SendData): Promise<TransferMessage | null> {
+  async send({ commit, dispatch }, payload: SendData): Promise<TransferMessage | null> {
     console.log('sending...', payload)
     commit(types.SET_SENDING, true)
     const { isNative, originNetwork, destNetwork, asset, amnt, recipient } =
@@ -199,6 +213,7 @@ const actions = <ActionTree<SDKState, RootState>>{
       commit(types.SET_SENDING, false)
       return transferMessage
     } catch (e) {
+      await dispatch('checkFailedHomes')
       console.error(e)
     }
 
@@ -241,18 +256,28 @@ const actions = <ActionTree<SDKState, RootState>>{
     }
     replica.connect(signer)
 
-    // prove and process
-    const receipt = await replica.proveAndProcess(
-      data.message as BytesLike,
-      data.proof.path,
-      data.proof.index
-    )
-    console.log('PROCESSED!!!!')
-    return receipt
+    try {
+      // prove and process
+      const receipt = await replica.proveAndProcess(
+        data.message as BytesLike,
+        data.proof.path,
+        data.proof.index
+      )
+      console.log('PROCESSED!!!!')
+      return receipt
+    } catch(e) {
+      await dispatch('checkFailedHomes')
+      console.error(e)
+    }
   },
 }
 
 const getters = <GetterTree<SDKState, RootState>>{
+  activeNetworks: (state: SDKState) => () => {
+    return Object.keys(networks)
+      .filter(n => !state.blacklist.has(networks[n].domainID))
+      .map(n => networks[n])
+  },
   getGasPrice: () => async (network: string | number) => {
     try {
       const provider = nomad.getProvider(network)
@@ -292,15 +317,7 @@ const getters = <GetterTree<SDKState, RootState>>{
 
   resolveRepresentation:
     () => async (network: string, token: TokenIdentifier) => {
-      let bridgeToken
-
-      try {
-        bridgeToken = await nomad.resolveRepresentation(network, token)
-      } catch (e) {
-        console.error(e)
-      }
-
-      return bridgeToken
+      return await nomad.resolveRepresentation(network, token)
     },
 }